add support for neo4j graphrag
