+++ src/langtrace_python_sdk/instrumentation/__init__.py
@@ -22,6 +22,7 @@ from .litellm import LiteLLMInstrumentation
 from .llamaindex import LlamaindexInstrumentation
 from .milvus import MilvusInstrumentation
 from .mistral import MistralInstrumentation
+from .neo4j_graphrag import Neo4jGraphRAGInstrumentation
 from .ollama import OllamaInstrumentor
 from .openai import OpenAIInstrumentation
 from .openai_agents import OpenAIAgentsInstrumentation
@@ -59,6 +60,7 @@ __all__ = [
     "AWSBedrockInstrumentation",
     "CerebrasInstrumentation",
     "MilvusInstrumentation",
+    "Neo4jGraphRAGInstrumentation",
     "GoogleGenaiInstrumentation",
     "CrewaiToolsInstrumentation",
     "GraphlitInstrumentation",
