+++ src/langtrace_python_sdk/instrumentation/neo4j_graphrag/__init__.py
@@ -1,0 +1,3 @@
+from .instrumentation import Neo4jGraphRAGInstrumentation
+
+__all__ = ["Neo4jGraphRAGInstrumentation"]
